Move smileyfunction to smileys component
